refactor(server): share SQLite + Postgres Database impls via a blanket impl
SQLite and Postgres speak the same `$1` placeholder dialect and, apart from one
table in `delete_user`, run byte-for-byte identical SQL. Rather than keep two
copies in sync, each now implements a small private `SqlxBackend` trait (a pool
accessor plus one hook) and picks up a full `Database` impl for free from a
blanket `impl<T: SqlxBackend> Database for T` in `db::shared`, where the shared
query bodies live as ordinary, navigable code.

- `Database` stays object-safe (the associated `Db` type lives on `SqlxBackend`,
  not `Database`), so `Arc<dyn Database>` is unaffected.
- The blanket impl is generic over the backend, so it carries the one-off sqlx
  trait bounds (`Executor`, `IntoArguments`, `FromRow`/`Encode`/`Type` for the
  row and bind types). These stay contained: handlers use `Arc<dyn Database>`,
  never `T: Database`, so they never see the bounds.
- The one divergence — Postgres also clearing `total_history_count_user` in
  `delete_user` — is expressed with the `delete_user_extra_statements` hook.
  It returns full `&'static str` statements (not table names) so they satisfy
  `SqlSafeStr`; a runtime-formatted `String` would not.
- MySQL is left hand-written and does not implement `SqlxBackend`: it needs `?`
  placeholders and differs in `add_user` (`last_insert_id()`) and `add_records`
  (`on duplicate key update`).

`#[instrument(skip_all)]` is now applied uniformly, so `delete_store` and
`status` gain a tracing span they previously lacked on both backends.

SQLite db_story and lib tests pass; the duplicated query bodies are gone.
